requirements.txt
asyncpg == 0.25.0
discord.py == 2.0.1
python_dotenv == 0.19.2
discord.py-pagination == 0.0.5